8bit -> 16bit eeprom adress
bug preventing to read/write more than 128 eeprom locations

+++ avr/common/methods.h
@@ -17,8 +17,8 @@ int8_t getVSCPFrame( uint16_t *pvscpclass,
                         uint8_t *pSize,
                         uint8_t *pData );
 
-int readEEPROM( uint8_t addr );
-int writeEEPROM( uint8_t addr, uint8_t data );
+int readEEPROM( uint16_t addr );
+int writeEEPROM( uint16_t addr, uint8_t data );
 
 uint8_t vscp_readNicknamePermanent( void );
 void vscp_writeNicknamePermanent( uint8_t nickname );
